fastapi-server/requirements.txt
fastapi==0.115.9
uvicorn==0.34.0
httpx==0.28.1
aiohttp==3.11.16
requests==2.32.3
itsdangerous==2.2.0
python-dotenv==1.1.0
bcrypt
passlib==1.7.4
pydantic==2.11.3
pydantic-settings==2.8.1
pydantic-extra-types==2.10.3
SQLAlchemy==2.0.40
pandas==2.2.3
langchain==0.3.23
langgraph==0.3.29
langchain-upstage==0.6.0
langchain_community
chromadb==1.0.4
openai==1.72.0
tiktoken==0.9.0
tokenizers==0.19.1
PyMySQL==1.1.1
langsmith
prophet
mcp
pydantic[email]
